Rename view handler file
